Time: 139 ms (56.85%) | Memory: 111.7 MB (18.62%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.java
@@ -51,10 +51,10 @@ class MedianFinder {
     public double findMedian() {
         int size = lMaxHeap.size() + rMinHeap.size();
         if((size&1) == 1){
-            return (double)rMinHeap.peek();
+            return rMinHeap.peek();
         }
 
-        return (rMinHeap.peek() + lMaxHeap.peek()) / (double)2;
+        return (rMinHeap.peek() + lMaxHeap.peek()) / 2.0;
         
     }
 }
